support: Add separate mesa-demos package
This will allow us to expose glxgears + glxgears 32bit, which will be
invaluable for debugging any potential LSI or driver issues. We'll create
a new `lsi-exec` binary in LSI to support the full initial bootstrap routine
for then executing arbitrary commands, so that they can be used to execute
glxgears + glxinfo with the full expected environment for Steam.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>